requirements.txt
# ============================================================================
# CORE AIRFLOW
# ============================================================================
apache-airflow==2.7.3

# ============================================================================
# AIRFLOW PROVIDERS
# ============================================================================
# HTTP Provider (for API calls)
apache-airflow-providers-http==4.6.0

# Amazon Provider (for S3 integration)
apache-airflow-providers-amazon==8.11.0

# ============================================================================
# DATA PROCESSING
# ============================================================================
# Data manipulation and analysis
pandas==2.1.3
numpy==1.26.2

# ============================================================================
# AWS SDK
# ============================================================================
# AWS SDK for Python
boto3==1.29.7
botocore==1.32.7

# ============================================================================
# UTILITIES
# ============================================================================
# HTTP library (used by Airflow providers)
requests==2.31.0

# JSON handling (included in Python, but listed for clarity)
# json - standard library

# Date/time utilities (included in Python)
# datetime - standard library
